HT_24/requirements.txt
celery==5.3.6
django==5.0
djangorestframework==3.14.0
factory_boy==3.3.0
requests==2.31.0
